requirements.txt
# pygame comes preinstalled on Raspberry PI
#pygame==1.9.6
pyserial==3.4
